+++ Assets/LDtkUnity/Editor/PropertyDrawer/LDtkSpriteLayerAttributeDrawer.cs
@@ -1,0 +1,20 @@
+﻿using UnityEditor;
+using UnityEngine;
+
+namespace LDtkUnity.Editor
+{
+	[CustomPropertyDrawer(typeof(LDtkLayerMaskAttribute))]
+	public class LDtkLayerMaskAttributeDrawer : PropertyDrawer
+	{
+		public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
+		{
+			if (property.propertyType != SerializedPropertyType.Integer)
+			{
+				EditorGUI.PropertyField(position, property, label);
+				return;
+			}
+			
+			property.intValue = EditorGUI.LayerField(position, label, property.intValue);
+		}
+	}
+}