+++ Assets/LDtkUnity/Editor/PropertyDrawer/LDtkTagAttributeDrawer.cs
@@ -1,0 +1,20 @@
+using UnityEditor;
+using UnityEngine;
+
+namespace LDtkUnity.Editor
+{
+    [CustomPropertyDrawer(typeof(LDtkTagAttribute))]
+    public class LDtkTagAttributeDrawer : PropertyDrawer
+    {
+        public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
+        {
+            if (property.propertyType != SerializedPropertyType.String)
+            {
+                EditorGUI.PropertyField(position, property, label);
+                return;
+            }
+
+            property.stringValue = EditorGUI.TagField(position, label, property.stringValue);
+        }
+    }
+}