+++ Assets/LDtkUnity/Editor/CustomEditor/SectionDrawers/LDtkSectionDrawer.cs
@@ -156,7 +156,7 @@ namespace LDtkUnity.Editor
                 return;
             }
 
-            Debug.LogWarning("LDtk: Not allowed to assign an imported LDtk GameObject. It would have result in a recursive crash.");
+            Debug.LogWarning("LDtk: Not allowed to assign an imported LDtk GameObject. It would have resulted in a recursive crash.");
             prop.objectReferenceValue = null;
         }
     }