+++ Assets/LDtkUnity/Runtime/Tools/LDtkTagAttribute.cs
@@ -1,0 +1,11 @@
+using UnityEngine;
+using UnityEngine.Internal;
+
+namespace LDtkUnity
+{
+    [ExcludeFromDocs]
+    public class LDtkTagAttribute : PropertyAttribute
+    {
+        
+    }
+}