+++ Assets/LDtkUnity/CHANGELOG.md
@@ -1,7 +1,8 @@
 # 2.1.5
 ###### July XX, 2021
+- Added a new Tag and Layer field to the IntGridTile asset
+  - Allows splitting up tilemaps with customized tags/layers
 - Fixed all HelpURL broken links (was broken since 2.1.0)
-- 
 
 # 2.1.4
 ###### July 15, 2021