+++ Assets/LDtkUnity/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added a new Tag and Layer field to the IntGridTile asset
   - Allows splitting up tilemaps with customized tags/layers
 - Fixed all HelpURL broken links (was broken since 2.1.0)
+- Added a `Save Project` button to the LDtkIntGridTile inspector
 
 # 2.1.4
 ###### July 15, 2021