+++ Assets/LDtkUnity/CHANGELOG.md
@@ -1,9 +1,11 @@
 # 2.1.5
 ###### July XX, 2021
-- Added a new Tag and Layer field to the IntGridTile asset
-  - Allows splitting up tilemaps with customized tags/layers
+- Added numerous features to the LDtkIntGridTileInspector
+  - Added a new Tag and Layer field to the IntGridTile asset
+    - Allows splitting up tilemaps with customized tags/layers
+  - Added a `Save Project` button
+  - Added a safeguard to prevent lag/crashes when assigning imported LDtk projects to the GameObject Prefab field
 - Fixed all HelpURL broken links (was broken since 2.1.0)
-- Added a `Save Project` button to the LDtkIntGridTile inspector
 
 # 2.1.4
 ###### July 15, 2021