+++ Assets/LDtkUnity/CHANGELOG.md
@@ -1,5 +1,5 @@
 # 2.1.5
-###### July XX, 2021
+###### July 21, 2021
 - Added numerous features to the LDtkIntGridTileInspector
   - Added a new Tag and Layer field to the IntGridTile asset
     - Allows splitting up tilemaps with customized tags/layers